% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CENSUS_POPULATIONS}
\alias{CENSUS_POPULATIONS}
\title{Mapping from Census demographic categories to StriveTogether ones}
\format{
\subsection{CENSUS_POPULATIONS}{
}
}
\usage{
CENSUS_POPULATIONS
}
\description{
Mapping from Census demographic categories to StriveTogether ones
}
\keyword{datasets}
