% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_acs_employment.R
\name{fetch_acs_employment}
\alias{fetch_acs_employment}
\title{Pull one year of data about household income and populations by sex and race}
\usage{
fetch_acs_employment(.year, .state, .county, ...)
}
\arguments{
\item{.year}{\verb{<int>} the year to pull data}

\item{.state}{\verb{<int>} the two-digit FIPS code for the state}

\item{.county}{\verb{<chr>} the three-digit FIPS code for the county}

\item{...}{
  Arguments passed on to \code{\link[hercacstables:fetch_data]{hercacstables::fetch_data}}
  \describe{
    \item{\code{use_key}}{<lgl?> optional, should the query include a Census API key from the system environment. Defaults to \code{TRUE}}
  }}
}
\value{
\verb{<tbl>} a data frame with columns based upon \code{...}
}
\description{
These data will all come from the 5-year ACS.
This function fetches tract-level data for one county in one state
}
