% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_FOR_ACS_EMPLOYMENT}
\alias{GLOSSARY_FOR_ACS_EMPLOYMENT}
\title{Explain Census variables in terms of Sustainable Wage calculations}
\format{
\subsection{GLOSSARY_FOR_ACS_EMPLOYMENT}{

an object of class \code{tbl_df}, \code{tbl}, \code{data.frame}
with 21 rows and 4 columns
\describe{
\item{Group     }{\verb{<chr>} the table that the data come from, e.g. B01001}
\item{Index     }{\verb{<int>} the variable's row number in its table }
\item{Gender    }{\verb{<chr>} the gender that this row corresponds to}
\item{Population}{\verb{<chr>} the socioeconomic group that this variable corresponds to}
}
}
}
\usage{
GLOSSARY_FOR_ACS_EMPLOYMENT
}
\description{
Explain Census variables in terms of Sustainable Wage calculations
}
\keyword{datasets}
