% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_POPS_FOR_CENSUS}
\alias{GLOSSARY_OF_POPS_FOR_CENSUS}
\title{Mapping from Census demographic categories to StriveTogether ones}
\format{
\subsection{GLOSSARY_OF_POPS_FOR_CENSUS}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 12 rows and 2 columns.
\describe{
\item{Category}{\verb{<chr>} one of the OMB categories used by IPEDS}
\item{Population}{\verb{<chr>} one of the categories used by StriveTogether}
}
}
}
\usage{
GLOSSARY_OF_POPS_FOR_CENSUS
}
\description{
Mapping from Census demographic categories to StriveTogether ones
}
\keyword{datasets}
