% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_POPS_FOR_CIA}
\alias{GLOSSARY_OF_POPS_FOR_CIA}
\title{The literal row labels for different populations that StriveTogether wants}
\format{
\subsection{GLOSSARY_OF_POPS_FOR_CIA}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 72 rows and 3 columns.
\describe{
\item{CIA Year}{\verb{<int>} which year that Strive requested this combination of Gender and Population}
\item{Gender}{\verb{<chr>} either "All", "Male", or "Female}
\item{Population}{\verb{<chr>} some racial, ethnic, language-learner, and economic categories}
}
}
}
\usage{
GLOSSARY_OF_POPS_FOR_CIA
}
\description{
The literal row labels for different populations that StriveTogether wants
}
\keyword{datasets}
