% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_POPS_FOR_COMPS}
\alias{GLOSSARY_OF_POPS_FOR_COMPS}
\title{Group population and gender classes into three comparsions for visualization}
\format{
\subsection{GLOSSARY_OF_POPS_FOR_COMPS}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 13 rows and 3 columns.
\describe{
\item{Comparison}{\verb{<chr>} either "SES," "Gender," or "Race/Ethnicity"}
\item{Gender}{\verb{<chr>} currently, Strive just uses two, plus "All"}
\item{Population}{\verb{<chr>} a specific socioeconomic status, race/ethnicity, or gender}
}
}
}
\usage{
GLOSSARY_OF_POPS_FOR_COMPS
}
\description{
Group population and gender classes into three comparsions for visualization
}
\keyword{datasets}
