% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_POPS_FOR_IPEDS}
\alias{GLOSSARY_OF_POPS_FOR_IPEDS}
\title{Mapping from IPEDS populations to StriveTogether ones}
\format{
\subsection{GLOSSARY_OF_POPS_FOR_IPEDS}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 27 rows and 4 columns.
\describe{
\item{Sex}{\verb{<chr>} "all", "men", or "women", as per IPEDS}
\item{Gender}{\verb{<chr>} "All", "Female", or "Male", as per StriveTogether}
\item{Race/Ethnicity}{\verb{<chr>} one of the OMB categories used by IPEDS}
\item{Population}{\verb{<chr>} one of the categories used by StriveTogether}
}
}
}
\usage{
GLOSSARY_OF_POPS_FOR_IPEDS
}
\description{
Mapping from IPEDS populations to StriveTogether ones
}
\keyword{datasets}
