% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_POPS_FOR_WISEDASH}
\alias{GLOSSARY_OF_POPS_FOR_WISEDASH}
\title{Mapping between WISEDash Public's "GROUP_BY_VALUE" and StriveTogether's "population."}
\format{
\subsection{GLOSSARY_OF_POPS_FOR_WISEDASH}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 13 rows and 3 columns.
\describe{
\item{Population}{\verb{<chr>} a demographic group label used in the Civic Infrastructure Assessment}
\item{GROUP_BY_VALUE}{\verb{<chr>} the value found in published WISEDash data}
\item{Gender}{\verb{<chr>} Either "All", "Female", or "Male"}
}
}
}
\source{
\url{https://strivetogether.my.site.com/s/article/2023-Civic-Infrastructure-Assessment-resources}
}
\usage{
GLOSSARY_OF_POPS_FOR_WISEDASH
}
\description{
Mapping between WISEDash Public's "GROUP_BY_VALUE" and StriveTogether's "population."
}
\keyword{datasets}
