% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_POPS_FOR_WISEDASH_LOCAL}
\alias{GLOSSARY_OF_POPS_FOR_WISEDASH_LOCAL}
\title{A mapping from WISEDash Local labels to StriveTogether ones.}
\format{
\subsection{GLOSSARY_OF_POPS_FOR_WISEDASH_LOCAL}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 14 rows and 4 columns.
\describe{
\item{GROUP_BY}{\verb{<chr>} broader demographic concepts such as Race/Ethnicity, Gender, or SES.}
\item{GROUP_BY_VALUE}{\verb{<chr>} specific instances of groups such as Race/Ethnicity, Gender, or SES.}
\item{Gender}{\verb{<chr>} currently, Strive just uses two, plus "All"}
\item{Population}{\verb{<chr>} a specific socioeconomic status, race/ethnicity, or gender}
}
}
}
\usage{
GLOSSARY_OF_POPS_FOR_WISEDASH_LOCAL
}
\description{
A mapping from WISEDash Local labels to StriveTogether ones.
}
\keyword{datasets}
