% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPEDS_POPULATIONS}
\alias{IPEDS_POPULATIONS}
\title{Mapping from IPEDS populations to StriveTogether ones}
\format{
\subsection{IPEDS_POPULATIONS}{
}
}
\usage{
IPEDS_POPULATIONS
}
\description{
Mapping from IPEDS populations to StriveTogether ones
}
\keyword{datasets}
