% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_partnerships.R
\name{load_partnerships}
\alias{load_partnerships}
\title{Load JSON descriptions of StriveTogether partnerships}
\usage{
load_partnerships(.path)
}
\arguments{
\item{.path}{\verb{<chr>} the full path to the partnership}
}
\value{
\verb{<lst>} a named list of partnership information. See Details.
}
\description{
Each StriveTogether partnership calculates its CIA report in a slightly
different way.
Nevertheless, there are many features in common.
This function expects a JSON file with an array of objects, one per
partnership.
The objects contain the details necessary for computing CIA outcomes.
}
\details{
The actual structure of the output depends upon the JSON file.
The current expectation is that each object has the following fields.
\describe{
\item{name \verb{<chr>}}{this will be used as the object's name in the output list.}
\item{spreadsheet_id \verb{<chr>}}{the uuid for the Google Sheet that will hold the partnership's results}
\item{state \verb{<chr>}}{the two-letter postal code for the state that the partnership works in}
\item{state lds info \verb{<lst>}}{an object with whatever info the state-specific function needs to compute outcomes}
\item{ipeds unit ids \verb{<int[]>}}{the UNITIDs for each of the partnership's postsecondary collaborators}
}
}
