% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_growth_for_k8.R
\name{map_growth_for_k8}
\alias{map_growth_for_k8}
\title{Import and wrangle WISEDash Local data about students' growth on MAP reading}
\usage{
map_growth_for_k8(
  .filepath,
  .basename_prefix,
  .grade_pattern,
  .basename_suffix = ".*"
)
}
\arguments{
\item{.filepath}{string, one or more full paths to raw data files}

\item{.basename_prefix}{string, a regex to match the parts of file names that appear before the grade}

\item{.grade_pattern}{string, a regex to match the part of the file name containing the grade level}

\item{.basename_suffix}{string, a regex to match the parts of file names that appear after the grade}
}
\value{
a data frame with 9 columns
}
\description{
This function both reads a raw CSV output from WISEDash Local and wrangles it
into a CIA-friendly format. In particular, it uses the "Test Growth Result"
field to count the number of students who met or exceeded their growth goal.
\strong{CAVEAT} the file name(s) must contain grade information.
}
\details{
The returned value will have the following structure:

\describe{
\item{Grade}{\if{html}{\out{<chr>}} A zero-padded two-digit grade such as "03" or "08"}
\item{Gender}{\if{html}{\out{<chr>}} A gender identity like "All," "Female," "Male," or "Non-binary."}
\item{Race/Ethnicity}{\if{html}{\out{<chr>}} A racial or ethnic identity from \href{https://dpi.wi.gov/wise/data-elements/race}{WISEDash's options}}
\item{EL}{\if{html}{\out{<chr>}} Whether ("Yes") or not ("No") a student is an English Language Learner}
\item{Year}{\if{html}{\out{<chr>}} The school year in YYYY-YYYY format}
\item{N}{\if{html}{\out{<int>}} The number of children in a combination of Gender, Race, and EL that met their growth goals.}
\item{D}{\if{html}{\out{<int>}} The number of children in a combination of Gender, Race, and EL}
}
}
\seealso{
\code{\link[=read_wisedash_local]{read_wisedash_local()}}
}
