% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_percentiles_for_kg.R
\name{map_percentiles_for_kg}
\alias{map_percentiles_for_kg}
\title{Import and wrangle WISEDash Local data about Kindergartners' percentile scores on MAP reading}
\usage{
map_percentiles_for_kg(.filepath, .standard)
}
\arguments{
\item{.filepath}{string, one or more full paths to raw data files}

\item{.standard}{float, the minimum percentile to count as proficient. 0-100 scale.}
}
\value{
a data frame with 9 columns
}
\description{
This function both reads a raw CSV output from WISEDash Local and wrangles it
into a CIA-friendly format. In particular, it converts the percentile scores
into counts of Kindergarten readiness by considering any percentile above the
\code{.standard} argument to be "ready."
}
\details{
The returned value will have the following structure:

\describe{
\item{File Name}{\if{html}{\out{<chr>}} The \code{\link[=basename]{basename()}} of the .filepath argument}
\item{Gender}{\if{html}{\out{<chr>}} A gender identity like "All," "Female," "Male," or "Non-binary."}
\item{Race/Ethnicity}{\if{html}{\out{<chr>}} A racial or ethnic identity from \href{https://dpi.wi.gov/wise/data-elements/race}{WISEDash's options}}
\item{EL}{\if{html}{\out{<chr>}} Whether ("Yes") or not ("No") a student is an English Language Learner}
\item{School Year}{\if{html}{\out{<chr>}} The school year in YY-YY format}
\item{N}{\if{html}{\out{<int>}} The number of children in a combination of Gender, Race, and EL with percentile scores that exceeded the \code{.standard}}
\item{D}{\if{html}{\out{<int>}} The number of children in a combination of Gender, Race, and EL}
\item{Year}{\if{html}{\out{<chr>}} The school year in YYYY-YYYY format}
\item{Grade}{\if{html}{\out{<chr>}} Always "KG"}
}
}
\seealso{
\code{\link[=read_wisedash_local]{read_wisedash_local()}}
}
