% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OUTPUT_POPULATIONS}
\alias{OUTPUT_POPULATIONS}
\title{The literal row labels for different populations that StriveTogether wants}
\format{
\subsection{OUTPUT_POPULATIONS}{

\describe{
\item{Gender}{either "All", "Male", or "Female}
\item{Population}{some racial, ethnic, language-learner, and economic categories}
}
}
}
\usage{
OUTPUT_POPULATIONS
}
\description{
The literal row labels for different populations that StriveTogether wants
}
\keyword{datasets}
