% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POSTSECONDARY_BY_SEX_AND_RACE}
\alias{POSTSECONDARY_BY_SEX_AND_RACE}
\title{Variables from the US Census to describe postsecondary completion by sex and race/ethnicity}
\format{
\subsection{POSTSECONDARY_BY_SEX_AND_RACE}{
}
}
\usage{
POSTSECONDARY_BY_SEX_AND_RACE
}
\description{
Variables from the US Census to describe postsecondary completion by sex and race/ethnicity
}
\keyword{datasets}
