% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POSTSECONDARY_EDUCATION_LEVELS}
\alias{POSTSECONDARY_EDUCATION_LEVELS}
\title{Values of education levels from the US Census to describe postsecondary completion}
\format{
\subsection{POSTSECONDARY_EDUCATION_LEVELS}{
}
}
\usage{
POSTSECONDARY_EDUCATION_LEVELS
}
\description{
Values of education levels from the US Census to describe postsecondary completion
}
\keyword{datasets}
