% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisedash_local.R
\name{predicate_MAP_growth}
\alias{predicate_MAP_growth}
\title{Identify which students met or exceeded their MAP growth goals}
\usage{
predicate_MAP_growth(.test_growth_result)
}
\arguments{
\item{.test_growth_result}{\verb{<chr[]>} a column of verbal descriptions of MAP growth results}
}
\value{
\verb{<lgl[]>} \code{TRUE} when the input value es "Met Growth"
}
\description{
Identify which students met or exceeded their MAP growth goals
}
