% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisedash_local.R
\name{predicate_MAP_percentile}
\alias{predicate_MAP_percentile}
\title{Identify which students met or exceeded their MAP growth goals}
\usage{
predicate_MAP_percentile(.percentile)
}
\arguments{
\item{.percentile}{\verb{<chr[]>} a column of percentiles in \verb{[0, 100]}}
}
\value{
\verb{<lgl[]>} \code{TRUE} when the input value is greater than 33.3
}
\description{
Identify which students met or exceeded their MAP growth goals
}
