% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hs_completion.R
\name{read_hs_completion}
\alias{read_hs_completion}
\title{Parse HS Completion data from WISEDash Public into a CIA-friendly format}
\usage{
read_hs_completion(.files)
}
\arguments{
\item{.files}{a series of paths to csv files that contain HS graduation data from WISEDash public}
}
\value{
a data frame with some WISEDash Public columns and some CIA columns
}
\description{
Parse HS Completion data from WISEDash Public into a CIA-friendly format
}
