% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_map_growth.R
\name{read_map_growth}
\alias{read_map_growth}
\title{Import a csv file with data from WISEDash Local about map growth}
\usage{
read_map_growth(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> full paths to data files}
}
\value{
a data frame with the following fields
\describe{
\item{Student ID}{\if{html}{\out{<integer>}}}
\item{Name}{\if{html}{\out{<character>}}}
\item{Status}{\if{html}{\out{<character>}}}
\item{Current School}{\if{html}{\out{<character>}}}
\item{Curr. Grade}{\if{html}{\out{<character>}}}
\item{Gender}{\if{html}{\out{<character>}}}
\item{Race/Ethnicity}{\if{html}{\out{<character>}}}
\item{EL}{\if{html}{\out{<character>}}}
\item{SPED}{\if{html}{\out{<character>}}}
\item{Test Growth Result}{\if{html}{\out{<character>}}}
\item{Test Growth Period}{\if{html}{\out{<character>}}}
\item{Year}{\if{html}{\out{<character>}}}
\item{Test Growth Percentile}{\if{html}{\out{<double>}}}
}
}
\description{
Import a csv file with data from WISEDash Local about map growth
}
