% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_postsecondary_enrollment.R
\name{read_postsecondary_enrollment}
\alias{read_postsecondary_enrollment}
\title{Parse post-secondary enrollment data from WISEDash Public into a CIA-friendly format}
\usage{
read_postsecondary_enrollment(.files)
}
\arguments{
\item{.files}{a series of paths to csv files that contain PS Enrollment data from WISEDash public}
}
\value{
a data frame with some WISEDash Public columns and some CIA columns
}
\description{
Parse post-secondary enrollment data from WISEDash Public into a CIA-friendly format
}
