% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wisedash_public.R
\name{read_wisedash_public}
\alias{read_wisedash_public}
\title{Import data from WISEDash Public into a CIA-friendly format}
\usage{
read_wisedash_public(.files, .spec)
}
\arguments{
\item{.files}{an array of paths to csv files that contain data from WISEDash public}

\item{.spec}{a list of column specifications for \code{\link[readr:read_delim]{readr::read_csv()}}}
}
\value{
a data frame with some WISEDash Public columns and some CIA columns
}
\description{
Import data from WISEDash Public into a CIA-friendly format
}
