% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SEMI_JOINS_FOR_WISEDASH_LOCAL}
\alias{SEMI_JOINS_FOR_WISEDASH_LOCAL}
\title{A lookup table for mapping demographic information from WISEDash Local to StriveTogether formats}
\format{
\subsection{SEMI_JOINS_FOR_WISEDASH_LOCAL}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 36 rows and 6 columns.
\describe{
\item{Gender}{\verb{<chr>} a StriveTogether gender identity}
\item{Population}{\verb{<chr>} a StriveTogether race/ethnicity, language status, or economic status}
\item{Gender Pattern}{\verb{<chr>} a regular expression to match the correct genders for this gender-population combination}
\item{Race Pattern}{\verb{<chr>} a regular expression to match the correct races for this gender-population combination}
\item{EL Pattern}{\verb{<chr>} a regular expression to match the correct language status for this gender-population combination}
\item{SES Pattern}{\verb{<chr>} a regular expression to match the correct economic status for this gender-population combination}
}
}
}
\usage{
SEMI_JOINS_FOR_WISEDASH_LOCAL
}
\description{
A lookup table for mapping demographic information from WISEDash Local to StriveTogether formats
}
\keyword{datasets}
