% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH}
\alias{SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH}
\title{Column specification for data from WISEDash about student growth in MAP scores}
\format{
\subsection{SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH}{

A list of 13 name-datatype pairs
\describe{
\item{Student ID}{\verb{<int>}}
\item{Name}{\verb{<chr>}}
\item{Status}{\verb{<chr>}}
\item{Current School}{\verb{<chr>}}
\item{Curr. Grade}{\verb{<chr>}}
\item{Gender}{\verb{<chr>}}
\item{Race/Ethnicity}{\verb{<chr>}}
\item{EL}{\verb{<chr>}}
\item{SPED}{\verb{<chr>}}
\item{Test Growth Result}{\verb{<chr>}}
\item{Test Growth Period}{\verb{<chr>}}
\item{Year}{\verb{<chr>}}
\item{Test Growth Percentile}{\verb{<dbl>}}
}
}
}
\usage{
SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH
}
\description{
Column specification for data from WISEDash about student growth in MAP scores
}
\keyword{datasets}
