% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES}
\alias{SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES}
\title{Column specification for data from WISEDash about student MAP percentile scores}
\format{
\subsection{SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES}{

\describe{
\item{Student ID}{\if{html}{\out{<int>}}}
\item{Name}{\if{html}{\out{<chr>}}}
\item{Status}{\if{html}{\out{<chr>}}}
\item{Current School}{\if{html}{\out{<chr>}}}
\item{Curr. Grade}{\if{html}{\out{<chr>}}}
\item{Gender}{\if{html}{\out{<chr>}}}
\item{Race/Ethnicity}{\if{html}{\out{<chr>}}}
\item{EL}{\if{html}{\out{<chr>}}}
\item{SPED}{\if{html}{\out{<chr>}}}
\item{Percentile Score}{\if{html}{\out{<chr>}}}
\item{Percentile}{\if{html}{\out{<dbl>}}}
\item{Norm Lvl}{\if{html}{\out{<chr>}}}
\item{School Year}{\if{html}{\out{<character>}}}
}
}
}
\usage{
SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES
}
\description{
Column specification for data from WISEDash about student MAP percentile scores
}
\keyword{datasets}
