% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_WISEDASH_PS_ENROLLMENT}
\alias{SPEC_FOR_WISEDASH_PS_ENROLLMENT}
\title{Column specification for data from WISEDash about Postsecondary Enrollment}
\format{
\subsection{SPEC_FOR_WISEDASH_PS_ENROLLMENT}{

\describe{
\item{SCHOOL_YEAR}{\if{html}{\out{<chr>}} the school year, in YYYY-YY format}
\item{AGENCY_TYPE}{\if{html}{\out{<chr>}} helps distinguish between district-run and independent charter schools}
\item{CHARTER_IND}{\if{html}{\out{<chr>}} whether or not a school is a charter}
\item{CESA}{\if{html}{\out{<chr>}}, which CESA, if any, the school belongs to}
\item{COUNTY}{\if{html}{\out{<chr>}}, the county that the school is located in}
\item{DISTRICT_CODE}{\if{html}{\out{<chr>}} the four-number code for each district, or "[Statewide]"}
\item{SCHOOL_CODE}{\if{html}{\out{<chr>}} the six-digit code for a school, or "[Districtwide]"}
\item{GRADE_GROUP}{\if{html}{\out{<chr>}} elementary, middle, high, k-12, etc.}
\item{DISTRICT_NAME}{\if{html}{\out{<chr>}} a human-readable name for a district. Not consistent across years}
\item{SCHOOL_NAME}{\if{html}{\out{<chr>}} a human-readable name for a school. Not consistent across years}
\item{INITIAL_ENROLLMENT}{\if{html}{\out{<chr>}} the time lag between graduation and enrollment}
\item{INSTITUTION_LOCATION}{\if{html}{\out{<chr>}} in-state or out-of-state}
\item{INSTITUTION_LEVEL}{\if{html}{\out{<chr>}} 2-year or 4-year}
\item{INSTITUTION_TYPE}{\if{html}{\out{<chr>}} public or private}
\item{GROUP_BY}{\if{html}{\out{<chr>}} demographic categories like race or gender}
\item{GROUP_BY_VALUE}{\if{html}{\out{<chr>}} demographic groups like Asian or Trans}
\item{GROUP_COUNT}{\if{html}{\out{<int>}} the number of students who could have enrolled in postsecondary education}
\item{STUDENT_COUNT}{\if{html}{\out{<int>}} the number of students who enrolled in postsecondary education}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=Postsecondary}
}
\usage{
SPEC_FOR_WISEDASH_PS_ENROLLMENT
}
\description{
Column specification for data from WISEDash about Postsecondary Enrollment
}
\keyword{datasets}
