% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STRIVE_TOGETHER_FORWARD}
\alias{STRIVE_TOGETHER_FORWARD}
\title{Use this to semi-join Forward results and only get Strive-relevant data}
\format{
\subsection{STRIVE_TOGETHER_FORWARD}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 2 rows and 3 columns
\describe{
\item{TEST_GROUP}{\emph{<chr>} "Forward"}
\item{GRADE_LEVEL}{\emph{<int>} \code{3L} or \code{8L}}
\item{TEST_SUBJECT}{\emph{<chr>} "ELA" for 3rd grade and "Mathematics" for 8th}
}
}
}
\usage{
STRIVE_TOGETHER_FORWARD
}
\description{
Use this to semi-join Forward results and only get Strive-relevant data
}
\keyword{datasets}
