% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STRIVE_TOGETHER_IPEDS}
\alias{STRIVE_TOGETHER_IPEDS}
\title{IPEDS fields for calculating 150\% graduation counts}
\format{
\subsection{STRIVE_TOGETHER_IPEDS}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 2 rows and 3 columns.
\describe{
\item{Status}{\emph{<chr>} "Adjusted Cohort" "Completed"}
\item{Relative Time}{\emph{<chr>} NA "<=150\%"}
\item{Credential Earned}{\emph{<chr>} NA "Any"}
}
}
}
\usage{
STRIVE_TOGETHER_IPEDS
}
\description{
IPEDS fields for calculating 150\% graduation counts
}
\keyword{datasets}
