% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STRIVE_TOGETHER_POPULATIONS}
\alias{STRIVE_TOGETHER_POPULATIONS}
\title{Mapping between race labels from source data to StriveTogether's terms}
\format{
\subsection{STRIVE_TOGETHER_POPULATIONS}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 11 rows and 4 columns.
\describe{
\item{Strive}{\emph{<chr>} a Race/Ethnicity label used in the Civic Infrastructure Assessment}
\item{Census Race/Ethnicity}{\emph{<chr>} a Race/Ethnicity label used by the US Census}
\item{WISEDash Public}{\emph{<chr>} Race/Ethniity labels used in published WISEDash data}
\item{WISEDash Local}{\emph{<chr>} Race/Ethnicity labels used in WISEDash Local reports}
}
}
}
\source{
\url{https://strivetogether.my.site.com/s/article/2023-Civic-Infrastructure-Assessment-resources}
}
\usage{
STRIVE_TOGETHER_POPULATIONS
}
\description{
Mapping between race labels from source data to StriveTogether's terms
}
\keyword{datasets}
