% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STRIVE_TOGETHER_RACE_ETHNICITIES}
\alias{STRIVE_TOGETHER_RACE_ETHNICITIES}
\title{A lookup table for mapping from StriveTogether race/ethnicities to those used by the Census, WISEDash Public, and WISEDash Local}
\format{
\subsection{STRIVE_TOGETHER_RACE_ETHNICITIES}{

\describe{
\item{Strive}{\if{html}{\out{<fct>}} ordered population identities}
\item{Census Race/Ethnicity}{\if{html}{\out{<chr>}} Census-defined racial and ethnic identities}
\item{WISEDash Public}{\if{html}{\out{<chr>}} Racial and ethnic identities as they appear in WISEDash Public}
\item{WISEDash Local}{\if{html}{\out{<chr>}} Racial and ethnic identities as they appear in WISEDash Local}
}
}
}
\usage{
STRIVE_TOGETHER_RACE_ETHNICITIES
}
\description{
A lookup table for mapping from StriveTogether race/ethnicities to those used by the Census, WISEDash Public, and WISEDash Local
}
\keyword{datasets}
