% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_subtotals.R
\name{success_subtotals}
\alias{success_subtotals}
\title{Count up successes and failures based upon}
\usage{
success_subtotals(.x, .success_field, ...)
}
\arguments{
\item{.x}{a data frame with at least one true/false column and at least one demographic column}

\item{.success_field}{the column with true/false values that designate succes}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> names of columns to group by}
}
\value{
a new data frame with \code{...}, "N" (the count of successes), and "D" (the size of the cohort)
}
\description{
Count up successes and failures based upon
}
