% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_map_growth.R
\name{tabulate_map_growth}
\alias{tabulate_map_growth}
\title{Count how many students met MAP growth expectations, using custom groups}
\usage{
tabulate_map_growth(.x, ...)
}
\arguments{
\item{.x}{<data.frame> the MAP growth data to tabulate}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> names of columns to group by}
}
\value{
a data frame with the ... fields, "Met", "Students", and "Percent"
}
\description{
Count how many students met MAP growth expectations, using custom groups
}
