% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_map_percentile.R
\name{tabulate_map_percentile}
\alias{tabulate_map_percentile}
\title{Count how many students met MAP percentile expectations, using custom groups}
\usage{
tabulate_map_percentile(.x, ...)
}
\arguments{
\item{.x}{<data.frame> the MAP percentile data to tabulate}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> names of columns to group by}
}
\value{
a data frame with the ... fields, "N" (proficient), and "D"  (cohort)
}
\description{
Count how many students met MAP percentile expectations, using custom groups
}
