% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_wide_cia_format.R
\name{to_wide_cia_format}
\alias{to_wide_cia_format}
\title{Reformat a long data frame into the CIA format}
\usage{
to_wide_cia_format(.long_data, .year)
}
\arguments{
\item{.long_data}{\verb{<tbl>} with columns "Year", "Gender", "Population", "N", and "D".}

\item{.year}{the year of the CIA}
}
\value{
a tibble with columns grouped by year, numerator, and denominator
}
\description{
Reformat a long data frame into the CIA format
}
