% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisedash_local.R
\name{wisedash_local_to_strive}
\alias{wisedash_local_to_strive}
\title{Compute marginal totals for the CIA from student-level data from WISEDash Local}
\usage{
wisedash_local_to_strive(.DATA, .success_column, .pred, ...)
}
\arguments{
\item{.DATA}{\verb{<tbl>} something made by \code{\link[=read_wisedash_local]{read_wisedash_local()}}}

\item{.success_column}{\verb{<chr>} the column to examine for success}

\item{.pred}{\verb{<fun>} a predicate function for determining success}

\item{...}{\code{\link[rlang:dyn-dots]{dynamic-dots}} Any additional grouping columns (using \link[tidyselect:select_helpers]{tidyselect::select_helpers})}
}
\value{
a data frame with columns that will help with pasting into the CIA
}
\description{
Compute marginal totals for the CIA from student-level data from WISEDash Local
}
