% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WISEDASH_POPULATIONS}
\alias{WISEDASH_POPULATIONS}
\title{Mapping between WISEDash Public's "GROUP_BY_VALUE" and StriveTogether's "population."}
\format{
\subsection{WISEDASH_POPULATIONS}{

\describe{
\item{Population}{a demographic group label used in the Civic Infrastructure Assessment}
\item{GROUP_BY_VALUE}{the value found in published WISEDash data}
\item{Gender}{Either "All", "Female", or "Male"}
}
}
}
\source{
\url{https://strivetogether.my.site.com/s/article/2023-Civic-Infrastructure-Assessment-resources}
}
\usage{
WISEDASH_POPULATIONS
}
\description{
Mapping between WISEDash Public's "GROUP_BY_VALUE" and StriveTogether's "population."
}
\keyword{datasets}
