% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisedash_public_to_strive.R
\name{wisedash_public_to_strive}
\alias{wisedash_public_to_strive}
\title{Wrangle WISEDash Public data into a format ready for the CIA}
\usage{
wisedash_public_to_strive(.x, .outcome_name, .group_name = "GROUP_COUNT")
}
\arguments{
\item{.x}{data from WISEDash public, either HS graduation or PS enrollment}

\item{.outcome_name}{either "HS Graduation" or "Post-secondary Enrollment"}

\item{.group_name}{optional, defaults to "GROUP_COUNT" but HS Completion uses
"COHORT_COUNT"}
}
\value{
a data frame with columns "Outcome", "Year", "Gender", "Population",
"Role", and "Value"
}
\description{
Wrangle WISEDash Public data into a format ready for the CIA
}
