% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_acs_employment.R
\name{wrangle_acs_employment}
\alias{wrangle_acs_employment}
\title{Estimate the number of people living in a household with a family-sustaining income}
\usage{
wrangle_acs_employment(.raw_acs_employment)
}
\arguments{
\item{.raw_acs_employment}{\verb{<tbl>} output from \code{\link[=fetch_acs_employment]{fetch_acs_employment()}}}
}
\value{
\verb{<tbl>} a data frame with five columns
\describe{
\item{Year}{\verb{<int>} Which 5-year ACS estimate the data come from}
\item{Gender}{\verb{<chr>} "All," "Female," or "Male."}
\item{Population}{\verb{<chr>} The socioeconomic group that the row describes}
\item{N}{\verb{<dbl>} The estimated number of people with a family-sustaining income}
\item{D}{\verb{<int>} The estimated number of people in total.}
}
}
\description{
Estimate the number of people living in a household with a family-sustaining income
}
