% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_ipeds_graduates.R
\name{wrangle_ipeds_graduates}
\alias{wrangle_ipeds_graduates}
\title{Wrangle raw IPEDS graduation data for local schools into the local 150\% postsecondary completion rate}
\usage{
wrangle_ipeds_graduates(.DATA, ...)
}
\arguments{
\item{.DATA}{a data frame with GR data a la \code{\link[hercipeds:read_graduates]{hercipeds::read_graduates()}}}

\item{...}{
  Arguments passed on to \code{\link[dplyr:filter]{dplyr::filter}}
  \describe{
    \item{\code{.by}}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}
    \item{\code{.preserve}}{Relevant when the \code{.data} input is grouped. If \code{.preserve = FALSE} (the default), the grouping structure is recalculated based on the
resulting data, otherwise the grouping is kept as is.}
  }}
}
\value{
a data frame ready for \code{\link[=to_wide_cia_format]{to_wide_cia_format()}}
}
\description{
Wrangle raw IPEDS graduation data for local schools into the local 150\% postsecondary completion rate
}
