% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_ps_enrollment.R
\name{wrangle_ps_enrollment}
\alias{wrangle_ps_enrollment}
\title{Convert raw data read from WISEDash Public files into a StriveTogether-friendly format}
\usage{
wrangle_ps_enrollment(.ps_data, .district = "4620")
}
\arguments{
\item{.ps_data}{a data frame in the format returned by \code{\link[=read_wisedash_public]{read_wisedash_public()}}}

\item{.district}{optional, the code for one district, defaults to Racine Unified's code.}
}
\value{
a tibble with six columns: Outcome, Year, Gender, Population, Role, and Value
\describe{
\item{SCHOOL_YEAR}{\if{html}{\out{<chr>}} the school year in YYYY-YYYY format}
\item{Population}{\if{html}{\out{<chr>}} a demographic category defined by StriveTogether}
\item{Gender}{\if{html}{\out{<chr>}} a gender identity recognized by StriveTogether}
\item{Postsecondary Enrollees}{\if{html}{\out{<int>}} the number of students who have ever enrolled in a postsecondary program}
\item{HS Completers}{\if{html}{\out{<int>}} the number of students who enrolled in a postsecondary program in the first fall after they graduated}
}
}
\description{
Convert raw data read from WISEDash Public files into a StriveTogether-friendly format
}
\seealso{
\code{\link[=read_wisedash_public]{read_wisedash_public()}}
}
