% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CENSUS_POPULATIONS}
\alias{CENSUS_POPULATIONS}
\title{Mapping from Census demographic categories to StriveTogether ones}
\format{
\subsection{CENSUS_POPULATIONS}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 12 rows and 2 columns.
\describe{
\item{Category}{\emph{<chr>} one of the OMB categories used by IPEDS}
\item{Population}{\emph{<chr>} one of the categories used by StriveTogether}
}
}
}
\usage{
CENSUS_POPULATIONS
}
\description{
Mapping from Census demographic categories to StriveTogether ones
}
\keyword{datasets}
