% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herccia-package.R
\docType{package}
\name{herccia-package}
\alias{herccia}
\alias{herccia-package}
\title{herccia: What the Package Does (One Line, Title Case)}
\description{
Perform the data tasks for a StriveTogether Civic Infrastructure Assessment
}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
