% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herccia-package.R
\docType{package}
\name{herccia-package}
\alias{herccia}
\alias{herccia-package}
\title{herccia: Utilities for creating Civic Infrastructure Assessment reports}
\description{
Perform the data tasks for a StriveTogether Civic Infrastructure Assessment. Initially, this package focuses on public data from the US Census and Wisconsin's Department of Public Instruction, and internal data from Racine Unified School District.
}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
