% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INCOME_BRACKETS}
\alias{INCOME_BRACKETS}
\title{Mapping from Census income brackets to dollar values}
\format{
\subsection{INCOME_BRACKETS}{
}
}
\usage{
INCOME_BRACKETS
}
\description{
Mapping from Census income brackets to dollar values
}
\keyword{datasets}
