% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JOIN_PATTERNS_FOR_WISEDASH_LOCAL}
\alias{JOIN_PATTERNS_FOR_WISEDASH_LOCAL}
\title{A lookup table for mapping demographic information from WISEDash Local to StriveTogether formats}
\format{
\subsection{JOIN_PATTERNS_FOR_WISEDASH_LOCAL}{

\describe{
\item{Gender}{\if{html}{\out{<chr>}} a StriveTogether gender identity}
\item{Population}{\if{html}{\out{<chr>}} a StriveTogether race/ethnicity, language status, or economic status}
\item{Gender Pattern}{a regular expression to match the correct genders for this gender-population combination}
\item{Race Pattern}{a regular expression to match the correct races for this gender-population combination}
\item{EL Pattern}{a regular expression to match the correct language status for this gender-population combination}
\item{SES Pattern}{a regular expression to match the correct economic status for this gender-population combination}
}
}
}
\usage{
JOIN_PATTERNS_FOR_WISEDASH_LOCAL
}
\description{
A lookup table for mapping demographic information from WISEDash Local to StriveTogether formats
}
\keyword{datasets}
