% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OUTPUT_POPULATIONS}
\alias{OUTPUT_POPULATIONS}
\title{The literal row labels for different populations that StriveTogether wants}
\format{
\subsection{OUTPUT_POPULATIONS}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 36 rows and 3 columns.
\describe{
\item{Gender}{\emph{<chr>} either "All", "Male", or "Female}
\item{Population}{\emph{<chr>} some racial, ethnic, language-learner, and economic categories}
\item{CIA Year}{\emph{<int>} which year that Strive requested this combination of Gender and Population}
}
}
}
\usage{
OUTPUT_POPULATIONS
}
\description{
The literal row labels for different populations that StriveTogether wants
}
\keyword{datasets}
