% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POSTSECONDARY_BY_LANGUAGE}
\alias{POSTSECONDARY_BY_LANGUAGE}
\title{Variables from the US Census to describe postsecondary completion by language spoken}
\format{
\subsection{POSTSECONDARY_BY_LANGUAGE}{
}
}
\usage{
POSTSECONDARY_BY_LANGUAGE
}
\description{
Variables from the US Census to describe postsecondary completion by language spoken
}
\keyword{datasets}
