% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POSTSECONDARY_BY_POVERTY}
\alias{POSTSECONDARY_BY_POVERTY}
\title{Variables from the US Census to describe postsecondary completion by poverty status}
\format{
\subsection{POSTSECONDARY_BY_POVERTY}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 12 rows and 6 columns.
\describe{
\item{Group      }{\verb{<chr>} the ACS table of this variable }
\item{Index      }{\verb{<int>} this variable's row in its ACS table}
\item{Variable   }{\verb{<chr>} the full specification of this variable for querying the API}
\item{Poverty Status}{\verb{<chr>} either "Above" or "Below"}
\item{Sex        }{\verb{<chr>} the sex of people counted by this variable}
\item{Status     }{\verb{<chr>} the educational status of people counted by this variable}
}
}
}
\usage{
POSTSECONDARY_BY_POVERTY
}
\description{
Variables from the US Census to describe postsecondary completion by poverty status
}
\keyword{datasets}
