% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POSTSECONDARY_BY_POVERTY}
\alias{POSTSECONDARY_BY_POVERTY}
\title{Variables from the US Census to describe postsecondary completion by poverty status}
\format{
\subsection{POSTSECONDARY_BY_POVERTY}{
}
}
\usage{
POSTSECONDARY_BY_POVERTY
}
\description{
Variables from the US Census to describe postsecondary completion by poverty status
}
\keyword{datasets}
