% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{POSTSECONDARY_BY_SEX_AND_RACE}
\alias{POSTSECONDARY_BY_SEX_AND_RACE}
\title{Variables from the US Census to describe postsecondary completion by sex and race/ethnicity}
\format{
\subsection{POSTSECONDARY_BY_SEX_AND_RACE}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 1428 rows and 7 columns.
\describe{
\item{Year       }{\verb{<int>} the ACS vintage that this variable belongs to}
\item{Group      }{\verb{<chr>} the ACS table of this variable }
\item{Index      }{\verb{<int>} this variable's row in its ACS table}
\item{Variable   }{\verb{<chr>} the full specification of this variable for querying the API}
\item{Sex        }{\verb{<chr>} the sex of people counted by this variable}
\item{Status     }{\verb{<chr>} the educational status of people counted by this variable}
\item{Census Race}{\verb{<chr>} the OMB race category of people counted by this variable}
}
}
}
\usage{
POSTSECONDARY_BY_SEX_AND_RACE
}
\description{
Variables from the US Census to describe postsecondary completion by sex and race/ethnicity
}
\keyword{datasets}
