% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wisedash_local.R
\name{read_wisedash_local}
\alias{read_wisedash_local}
\title{Import data from WISEDash Local into a CIA-friendly format}
\usage{
read_wisedash_local(.files, .spec, .success_column, .pred, ...)
}
\arguments{
\item{.files}{an array of paths to csv files that contain data from WISEDash Local}

\item{.spec}{a list of column specifications for \code{\link[readr:read_delim]{readr::read_csv()}}}

\item{.success_column}{the column to examine for success}

\item{.pred}{a predicate function for determining success}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Any additional grouping columns (using \link[tidyselect:language]{tidyselect::select_helpers})}
}
\value{
a data frame with some WISEDash Local columns and some CIA columns
}
\description{
Import data from WISEDash Local into a CIA-friendly format
}
