% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisedash_local.R
\name{read_wisedash_local}
\alias{read_wisedash_local}
\title{Import data from WISEDash Local into a CIA-friendly format}
\usage{
read_wisedash_local(.files, .spec)
}
\arguments{
\item{.files}{\verb{<chr[]>} an array of paths to csv files that contain data from WISEDash Local}

\item{.spec}{\verb{<chr[]>} a list of column specifications for \code{\link[readr:read_delim]{readr::read_csv()}}}
}
\value{
a data frame with columns defined by \code{.spec}
}
\description{
Import data from WISEDash Local into a CIA-friendly format
}
