% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH}
\alias{SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH}
\title{Column specification for data from WISEDash about student growth in MAP scores}
\format{
\subsection{SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH}{

\describe{
\item{Student ID}{\if{html}{\out{<integer>}}}
\item{Name}{\if{html}{\out{<character>}}}
\item{Status}{\if{html}{\out{<character>}}}
\item{Current School}{\if{html}{\out{<character>}}}
\item{Curr. Grade}{\if{html}{\out{<character>}}}
\item{Gender}{\if{html}{\out{<character>}}}
\item{Race/Ethnicity}{\if{html}{\out{<character>}}}
\item{EL}{\if{html}{\out{<character>}}}
\item{SPED}{\if{html}{\out{<character>}}}
\item{Test Growth Result}{\if{html}{\out{<character>}}}
\item{Test Growth Period}{\if{html}{\out{<character>}}}
\item{Year}{\if{html}{\out{<character>}}}
\item{Test Growth Percentile}{\if{html}{\out{<double>}}}
}
}
}
\usage{
SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH
}
\description{
Column specification for data from WISEDash about student growth in MAP scores
}
\keyword{datasets}
