% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH}
\alias{SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH}
\title{Column specification for data from WISEDash about student growth in MAP scores}
\format{
\subsection{SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH}{

A list of 13 name-datatype pairs
\describe{
\item{Student ID}{\emph{<int>}}
\item{Name}{\emph{<chr>}}
\item{Status}{\emph{<chr>}}
\item{Current School}{\emph{<chr>}}
\item{Curr. Grade}{\emph{<chr>}}
\item{Gender}{\emph{<chr>}}
\item{Race/Ethnicity}{\emph{<chr>}}
\item{EL}{\emph{<chr>}}
\item{SPED}{\emph{<chr>}}
\item{Test Growth Result}{\emph{<chr>}}
\item{Test Growth Period}{\emph{<chr>}}
\item{Year}{\emph{<chr>}}
\item{Test Growth Percentile}{\emph{<dbl>}}
}
}
}
\usage{
SPEC_FOR_WISEDASH_LOCAL_MAP_GROWTH
}
\description{
Column specification for data from WISEDash about student growth in MAP scores
}
\keyword{datasets}
