% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES}
\alias{SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES}
\title{Column specification for data from WISEDash about student MAP percentile scores}
\format{
\subsection{SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES}{

A list of 16 name-datatype pairs
\describe{
\item{Student ID}{\emph{<int>}}
\item{Name}{\emph{<chr>}}
\item{Status}{\emph{<chr>}}
\item{Current School}{\emph{<chr>}}
\item{Curr. Grade}{\emph{<chr>}}
\item{Gender}{\emph{<chr>}}
\item{Race/Ethnicity}{\emph{<chr>}}
\item{EL}{\emph{<chr>}}
\item{SPED}{\emph{<chr>}}
\item{Percentile Score}{\emph{<chr>}}
\item{Percentile}{\emph{<dbl>}}
\item{Norm Lvl}{\emph{<chr>}}
\item{School Year}{\emph{<chr>}}
}
}
}
\usage{
SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES
}
\description{
Column specification for data from WISEDash about student MAP percentile scores
}
\keyword{datasets}
