% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES}
\alias{SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES}
\title{Column specification for data from WISEDash about student MAP percentile scores}
\format{
\subsection{SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES}{

A list of 16 name-datatype pairs
\describe{
\item{Student ID}{\verb{<int>}}
\item{Name}{\verb{<chr>}}
\item{Status}{\verb{<chr>}}
\item{Current School}{\verb{<chr>}}
\item{Curr. Grade}{\verb{<chr>}}
\item{Gender}{\verb{<chr>}}
\item{Race/Ethnicity}{\verb{<chr>}}
\item{EL}{\verb{<chr>}}
\item{SPED}{\verb{<chr>}}
\item{Percentile Score}{\verb{<chr>}}
\item{Percentile}{\verb{<dbl>}}
\item{Norm Lvl}{\verb{<chr>}}
\item{School Year}{\verb{<chr>}}
}
}
}
\usage{
SPEC_FOR_WISEDASH_LOCAL_MAP_PERCENTILES
}
\description{
Column specification for data from WISEDash about student MAP percentile scores
}
\keyword{datasets}
