% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STRIVE_TOGETHER_RACE_ETHNICITIES}
\alias{STRIVE_TOGETHER_RACE_ETHNICITIES}
\title{A lookup table for mapping from StriveTogether race/ethnicities to those used by the Census, WISEDash Public, and WISEDash Local}
\format{
\subsection{STRIVE_TOGETHER_RACE_ETHNICITIES}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 8 rows and 2 columns.
\describe{
\item{Strive}{\emph{<chr>} ordered population identities}
\item{WISEDash Local}{\emph{<chr>} Racial and ethnic identities as they appear in WISEDash Local}
}
}
}
\usage{
STRIVE_TOGETHER_RACE_ETHNICITIES
}
\description{
A lookup table for mapping from StriveTogether race/ethnicities to those used by the Census, WISEDash Public, and WISEDash Local
}
\keyword{datasets}
