% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisedash_local_to_strive.R
\name{wisedash_local_to_strive}
\alias{wisedash_local_to_strive}
\title{Compute marginal totals for the CIA from student-level data from WISEDash Local}
\usage{
wisedash_local_to_strive(
  .x,
  .grade = "Grade",
  .year = "Year",
  .race = "Race/Ethnicity",
  .gender = "Gender",
  .ell = "EL"
)
}
\arguments{
\item{.x}{<data.frame> the student-level data}

\item{.grade, .year, .race, .gender, .ell}{\if{html}{\out{<chr>}} optional, the names of columns in \code{.x}}
}
\value{
a data frame with columns that will help with pasting into the CIA
}
\description{
Compute marginal totals for the CIA from student-level data from WISEDash Local
}
