% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisedash_local.R
\name{wrangle_wisedash_local}
\alias{wrangle_wisedash_local}
\title{Wrangle data from WISEDash Local into a CIA-friendly format}
\usage{
wrangle_wisedash_local(.DATA, .success_column, .pred, ...)
}
\arguments{
\item{.DATA}{\verb{<tbl>} something made by \code{\link[=read_wisedash_local]{read_wisedash_local()}}}

\item{.success_column}{\verb{<chr>} the column to examine for success}

\item{.pred}{\verb{<fun>} a predicate function for determining success}

\item{...}{\code{\link[rlang:dyn-dots]{dynamic-dots}} Any additional grouping columns (using \link[tidyselect:language]{tidyselect::select_helpers})}
}
\value{
a data frame with some WISEDash Local columns and some CIA columns
}
\description{
Wrangle data from WISEDash Local into a CIA-friendly format
}
